migrate to graphql
trpc is nice but becomes problematic for some things (i wanted to put images in a separate table with some metadata, but trpc in rust with sqlx makes it hard to load relations like that)

also simplifies the database (removes season types, which removes the need for a lot of recursive queries). not sure if this is a good move long-term but for now its fine.